+++ test/beast/core/Jamfile
@@ -33,6 +33,7 @@ local SOURCES =
     dynamic_buffer_ref.cpp
     error.cpp
     file.cpp
+    file_base.cpp
     file_posix.cpp
     file_stdio.cpp
     file_win32.cpp
